Fixed implementation of calculation of when to re-route. Also polish.

+++ src/main/java/org/appland/settlers/model/GameUtils.java
@@ -341,7 +341,7 @@ public class GameUtils {
         double             bestCaseCost;
 
         /* Define starting parameters */
-        bestCaseCost = (double)getDistanceInGameSteps(start, goal);
+        bestCaseCost = getDistanceInGameSteps(start, goal);
         toEvaluate.add(start);
         realCostToPoint.put(start, (double)0);
         estimatedFullCost.put(start, bestCaseCost);
